db: stormwind npc positions
Added what I believe to be all the stormwind npc locations that were changed, some could be missing.

Please keep in mind I did not have time to test this code fully and will not until next week. It shouldnt have any errors as it was tested in my other branch but I copy pasted it by hand to my main as I am not around my addon device.