End of my Adventure
There is too much to list but this contains all quests currently known aside from a couple with blank database pages. There are also about 155 that are missing some info but that is because the database does not contain the cords those quests however are documented in the google spreadsheet so others can find those and contribute.